Get real sizes in one batch
This is better than the depth-first approach. If one of the filenames
doesn't exist, the user will find out immediately. Before, if they
supplied 1,000 files, they might get through several hundred before
erroring out. Now, the error happens right away, and the problem can be
fixed.

This may ignore the situation where the file exists when we get its size
but doesn't exist by the time we're finished executing. But that kind of
thing isn't really within the purview of this program. It just tells you
sizes.